Fix issue where department updates didnt refresh
