--- app/admin/departments/create/page.tsx
@@ -1,172 +1,0 @@
-"use client";
-
-import type React from "react";
-
-import { useState } from "react";
-import { useRouter } from "next/navigation";
-import Link from "next/link";
-import { ArrowLeft, Save } from "lucide-react";
-
-import { Button } from "@/components/ui/button";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import { Textarea } from "@/components/ui/textarea";
-import { toast } from "sonner";
-// Mock teachers
-const teachers = [
-  { id: "1", name: "John Smith" },
-  { id: "2", name: "Sarah Johnson" },
-  { id: "3", name: "Michael Brown" },
-  { id: "4", name: "Emily Davis" },
-  { id: "5", name: "David Wilson" },
-];
-
-export default function CreateDepartmentPage() {
-  const router = useRouter();
-
-  const [formData, setFormData] = useState({
-    name: "",
-    description: "",
-    headOfDepartmentId: "",
-    established: "",
-  });
-
-  const [isSubmitting, setIsSubmitting] = useState(false);
-
-  const handleChange = (
-    e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-  ) => {
-    const { name, value } = e.target;
-    setFormData((prev) => ({ ...prev, [name]: value }));
-  };
-
-  const handleHeadOfDepartmentChange = (value: string) => {
-    setFormData((prev) => ({ ...prev, headOfDepartmentId: value }));
-  };
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsSubmitting(true);
-
-    // Simulate API call
-    setTimeout(() => {
-      toast.success("Department created successfully");
-      setIsSubmitting(false);
-      router.push("/admin/departments");
-    }, 1000);
-  };
-
-  return (
-    <div className="space-y-6">
-      <div className="flex items-center gap-2">
-        <Button variant="ghost" size="icon" asChild>
-          <Link href="/admin/departments">
-            <ArrowLeft className="h-4 w-4" />
-            <span className="sr-only">Back</span>
-          </Link>
-        </Button>
-        <div>
-          <h1 className="text-3xl font-bold tracking-tight dark:text-white">
-            Add New Department
-          </h1>
-          <p className="text-muted-foreground dark:text-gray-400">
-            Create a new department
-          </p>
-        </div>
-      </div>
-
-      <form onSubmit={handleSubmit}>
-        <Card>
-          <CardHeader>
-            <CardTitle>Department Information</CardTitle>
-            <CardDescription>Enter the department details</CardDescription>
-          </CardHeader>
-          <CardContent className="space-y-6">
-            <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-              <div className="space-y-2">
-                <Label htmlFor="name">Department Name</Label>
-                <Input
-                  id="name"
-                  name="name"
-                  value={formData.name}
-                  onChange={handleChange}
-                  required
-                  placeholder="e.g. Computer Science"
-                />
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="established">Established Year</Label>
-                <Input
-                  id="established"
-                  name="established"
-                  value={formData.established}
-                  onChange={handleChange}
-                  required
-                  placeholder="e.g. 1985"
-                />
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="headOfDepartment">Head of Department</Label>
-                <Select
-                  value={formData.headOfDepartmentId}
-                  onValueChange={handleHeadOfDepartmentChange}
-                >
-                  <SelectTrigger id="headOfDepartment">
-                    <SelectValue placeholder="Select head of department" />
-                  </SelectTrigger>
-                  <SelectContent>
-                    {teachers.map((teacher) => (
-                      <SelectItem key={teacher.id} value={teacher.id}>
-                        {teacher.name}
-                      </SelectItem>
-                    ))}
-                  </SelectContent>
-                </Select>
-              </div>
-            </div>
-            <div className="space-y-2">
-              <Label htmlFor="description">Description</Label>
-              <Textarea
-                id="description"
-                name="description"
-                value={formData.description}
-                onChange={handleChange}
-                rows={4}
-                placeholder="Enter department description..."
-              />
-            </div>
-            <div className="flex justify-end gap-3">
-              <Button variant="outline" asChild>
-                <Link href="/admin/departments">Cancel</Link>
-              </Button>
-              <Button type="submit" disabled={isSubmitting}>
-                {isSubmitting ? (
-                  "Creating..."
-                ) : (
-                  <>
-                    <Save className="h-4 w-4 mr-2" />
-                    Create Department
-                  </>
-                )}
-              </Button>
-            </div>
-          </CardContent>
-        </Card>
-      </form>
-    </div>
-  );
-}
